feat(agent): add agent command group absorbing cliproxy setup
New agent CLI command group with shared setup core relocated to a neutral module (agent/setup-core/); cliproxy setup and its helpers become thin compatibility wrappers with zero behavior change. agent setup provides model-credential parity; S3 wiring and workflow verifier land in later units. MCP allowlist unchanged (mutating command, CLI-only).
