+++ packages/cli/src/commands/agent/index.ts
@@ -1,0 +1,7 @@
+import type {goke} from 'goke'
+
+import {registerAgentSetup} from './setup'
+
+export function registerAgentCommands(cli: ReturnType<typeof goke>): void {
+  registerAgentSetup(cli)
+}
